temp addition to cardmaker.clj
Was able to use NetBeans to create a sample Java program, NumberAddition.
Added  to :import clause of ns
statement. From REPL, executed this ns statement, then was able to invoke
the java program; no code other than this one line was added to any
clj files in this project.

In case it was not said elsewhere, make-icard now makes a new infocard
from within Clojure.